feat: move NIP-05 display to profile card footer
- Remove NIP-05 display from AuthorBadge component
- Create new Nip05Display component with all NIP-05 logic
- Add NIP-05 display to ProfileCard footer for all profile views
- Maintains all existing functionality (verification, root detection, icons)
- NIP-05 now appears in footer for both individual profiles and profile lists
